Tests: Improve test organization (#4798)
* Improve test organization

* Add reload action to trigger local storage changes in ci

* Update swap string
