+++ apps/web/cypress/e2e/regression/create_safe_simple.cy.js
@@ -122,6 +122,7 @@ describe('Safe creation tests', () => {
         main.addToLocalStorage(constants.localStorageKeys.SAFE_v2__addressBook, ls.addressBookData.sameOwnerName),
       )
       .then(() => {
+        cy.reload()
         createwallet.waitForConnectionMsgDisappear()
         createwallet.selectMultiNetwork(1, constants.networks.sepolia.toLowerCase())
         createwallet.clickOnNextBtn()
