+++ apps/web/cypress/e2e/smoke/load_safe.cy.js
@@ -8,7 +8,6 @@ import { getSafes, CATEGORIES } from '../../support/safes/safesHandler.js'
 let staticSafes = []
 
 const testSafeName = 'Test safe name'
-const testOwnerName = 'Test Owner Name'
 
 describe('[SMOKE] Load Safe tests', () => {
   before(async () => {
@@ -17,14 +16,6 @@ describe('[SMOKE] Load Safe tests', () => {
 
   beforeEach(() => {
     cy.visit(constants.loadNewSafeSepoliaUrl)
-  })
-
-  it('[SMOKE] Verify a network can be selected in the Safe', () => {
-    safe.clickNetworkSelector(constants.networks.sepolia)
-    safe.selectPolygon()
-    cy.wait(2000)
-    safe.clickNetworkSelector(constants.networks.polygon)
-    safe.selectSepolia()
   })
 
   it('[SMOKE] Verify only valid Safe name can be accepted', () => {
@@ -37,7 +28,6 @@ describe('[SMOKE] Load Safe tests', () => {
 
     safe.verifyAddressInputValue(staticSafes.SEP_STATIC_SAFE_4)
     safe.verifyNextButtonStatus('be.enabled')
-    safe.clickOnNextBtn()
   })
 
   it('[SMOKE] Verify names cannot have more than 50 characters', () => {
@@ -50,44 +40,12 @@ describe('[SMOKE] Load Safe tests', () => {
     safe.inputAddress(constants.ENS_TEST_SEPOLIA)
     safe.verifyAddressInputValue(staticSafes.SEP_STATIC_SAFE_6)
     safe.verifyNextButtonStatus('be.enabled')
-    safe.clickOnNextBtn()
-  })
-
-  it('[SMOKE] Verify the custom Safe name is successfully loaded', () => {
-    safe.inputNameAndAddress(testSafeName, staticSafes.SEP_STATIC_SAFE_3)
-    safe.clickOnNextBtn()
-    createwallet.typeOwnerName(testOwnerName, 0)
-    safe.clickOnNextBtn()
-    safe.verifyDataInReviewSection(
-      testSafeName,
-      testOwnerName,
-      constants.commonThresholds.oneOfOne,
-      constants.networks.sepolia,
-      constants.SEPOLIA_OWNER_2,
-    )
-    safe.clickOnAddBtn()
-    main.verifyHomeSafeUrl(staticSafes.SEP_STATIC_SAFE_3)
-    safe.veriySidebarSafeNameIsVisible(testSafeName)
-    safe.verifyOwnerNamePresentInSettings(testOwnerName)
   })
 
   it('[SMOKE] Verify safe name has a default name', () => {
     createwallet.verifyDefaultWalletName(createwallet.defaultSepoliaPlaceholder)
     cy.reload()
     createwallet.verifyDefaultWalletName(createwallet.defaultSepoliaPlaceholder)
-  })
-
-  it('[SMOKE] Verify that after loading existing Safe, its name input is not empty', () => {
-    safe.inputNameAndAddress(testSafeName, staticSafes.SEP_STATIC_SAFE_4)
-    safe.clickOnNextBtn()
-    safe.verifyOnwerInputIsNotEmpty(0)
-  })
-
-  it('[SMOKE] Verify that when changing a network in dropdown, the same network is displayed in right top corner', () => {
-    safe.clickNetworkSelector(constants.networks.sepolia)
-    safe.selectPolygon()
-    cy.wait(1000)
-    safe.checkMainNetworkSelected(constants.networks.polygon)
   })
 
   it('[SMOKE] Verify there are mandatory networks in dropdown: Eth, Polygon, Sepolia', () => {
