+++ apps/web/cypress/e2e/smoke/tx_history.cy.js
@@ -88,11 +88,4 @@ describe('[SMOKE] Tx history tests', () => {
     )
     createTx.verifySpamIconIsDisplayed(typeUntrustedToken.title, typeUntrustedToken.summaryTxInfo)
   })
-
-  it('[SMOKE] Verify that copying sender address of untrusted token shows warning popup', () => {
-    createTx.toggleUntrustedTxs()
-    createTx.clickOnTransactionItemByName(typeUntrustedToken.summaryTitle, typeUntrustedToken.summaryTxInfo)
-    createTx.clickOnCopyBtn(0)
-    createTx.verifyWarningModalVisible()
-  })
 })
